Nov PacientApiController

+++ eZdravje/Controllers/PatientsController.cs
@@ -351,7 +351,12 @@ namespace eZdravje.Controllers
         {
             var patient = await _context.Patients.FindAsync(id);
             var activationCode = await _context.ActivationCodes.Where(x => x.UserId == id.ToString()).FirstOrDefaultAsync();
-            _context.ActivationCodes.Remove(activationCode);
+            
+            if (activationCode != null)
+            {
+                _context.ActivationCodes.Remove(activationCode);
+            }
+
             _context.Patients.Remove(patient);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
